Create Minesweeper
Just the code that I worked on, I couldn't get the machine learning part to work with keras, so the working part is basically a web scraper that opens a minesweeper game and is able to read the game's state. Also, selenium, which was used, updated, so the code used to open chrome here is outdated.